refactor(js/ai): moved generate api into ai package to enable use with middleware
